+++ frameworks/imagine_after_move.py
@@ -89,11 +89,11 @@ def imagineAfterMove_data(batch_size):
     texts = prompts_imagineAfterMove_tensor[inds]
     
     instructions = move_instructions_imagineAfterMove_tensor[inds]
-    imgs_out = torch.zeros(batch_size, 224, 224, 3, dtype=torch.bfloat16)
+    imgs_out = torch.zeros(batch_size, 224, 224, 3, dtype=torch.float32)
     for i in range(batch_size):
         G2 = discreteGame(S[i])
         process_steps(G2, instructions[i])
-        imgs_out[i] = torch.tensor(G2.getData(), dtype=torch.bfloat16)
+        imgs_out[i] = torch.tensor(G2.getData(), dtype=torch.float32)
     imgs_out = torch.permute(imgs_out, (0, 3, 1, 2)).contiguous().to(device)
 
     return texts, imgs_in, imgs_out
